Further refactoring of cmdfw.
Replace duplication of structures with duplication of methods at the
expense of some runtime overhead. Code of the methods does not
need to be modified when command line options are added, so now the
clients' code should be easier to maintain. I don't care about the
little runtime overhead I introduced as command-line parsing is not
performance-critical in this app.